requirements.txt
# AI Dev Team — Python dependencies
# Install: pip install -r requirements.txt

pyyaml>=6.0
